==> 10_code/notebook_1.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import statsmodels.formula.api as smf
from zipfile import ZipFile

# Load the data

DATA_PATH = "../00_source_data"
arrests_data = os.path.join(DATA_PATH, 'ucr_arrests_monthly_all'
                                       '_crimes_race_sex_1974_2020_dta.zip')

zip_file = ZipFile(arrests_data)
dfs = {text_file.filename: pd.read_stata(zip_file.open(text_file.filename))
       for text_file in zip_file.infolist()
       if text_file.filename.endswith('.dta')}
zip_file.close()

print(dfs.keys())
arrests_2019 = dfs['ucr_arrests_monthly_all_crimes_race_sex_2019.dta'].copy()
arrests_2020 = dfs['ucr_arrests_monthly_all_crimes_race_sex_2020.dta'].copy()
print(arrests_2019.shape)
print(arrests_2020.shape)

zip_file = ZipFile(os.path.join(DATA_PATH, 'crime.zip'))
dfs = {text_file.filename: pd.read_csv(zip_file.open(text_file.filename))
       for text_file in zip_file.infolist()
       if text_file.filename.endswith('.csv')}
zip_file.close()

print(dfs.keys())
crime_data = dfs['crime.csv'].copy()
offense_codes = dfs['offense_codes.csv'].copy()
print(crime_data.shape)
print(offense_codes.shape)

# Preprocessing

arrests_2019_cols = set(arrests_2019.columns)
arrests_2020_cols = set(arrests_2020.columns)
print(arrests_2019_cols.difference(arrests_2020_cols))
print("*" * 50)
print(arrests_2020_cols.difference(arrests_2019_cols))
print("*" * 50)
print(len(arrests_2020_cols.intersection(arrests_2019_cols)))

columns_to_remove = list(arrests_2020_cols.difference(arrests_2019_cols))
arrests_2020.drop(columns_to_remove, axis=1, inplace=True)
assert arrests_2019.shape[1] == arrests_2020.shape[1]

arrests_2020.head()

arrests_2020_filtered = arrests_2020.query("state_abb == 'CO'").copy()
arrests_2019_filtered = arrests_2019.query("state_abb == 'CO'").copy()

arrests_2020.loc[(arrests_2020['fips_state_county_code'] == '') &
                 (arrests_2020['agency_name'] == 'us secret service, denve'), 'fips_state_county_code'] = '08031'

arrests_2019.loc[(arrests_2019['fips_state_county_code'] == '') &
                 (arrests_2019['agency_name'] == 'us secret service, denve'), 'fips_state_county_code'] = '08031'

arrests_2020_filtered = arrests_2020_filtered.query("fips_state_county_code in ['08031', '08013', '08014']").copy()
arrests_2019_filtered = arrests_2019_filtered.query("fips_state_county_code in ['08031', '08013', '08014']").copy()
print(arrests_2020_filtered.shape)
print(arrests_2019_filtered.shape)

general_cols = arrests_2020_filtered.columns.to_list()[:14]
arrest_cols = [col for col in arrests_2020_filtered.columns if 'tot_arrest' in col]
arrests_2020_filtered = arrests_2020_filtered[general_cols + arrest_cols]
arrests_2019_filtered = arrests_2019_filtered[general_cols + arrest_cols]

arrests_df = pd.concat([arrests_2019_filtered, arrests_2020_filtered], axis=0)
print(arrests_df.shape)

arrests_df.head()

arrests_agg = (arrests_df
               .groupby(['month', 'year', 'fips_state_county_code'], as_index=False)[arrest_cols].sum())

population_norm = (arrests_df
                   .groupby(['month', 'year', 'fips_state_county_code', 'fips_place_code'], as_index=False)[
                       'population']
                   .max()
                   .groupby(['month', 'year', 'fips_state_county_code'], as_index=False)[['population']]
                   .sum())

merged = pd.merge(arrests_agg, population_norm, on=['month', 'year', 'fips_state_county_code'])
print(merged.shape)

merged.head()

violent_crime = ['agg_assault_tot_arrests', 'arson_tot_arrests', 'burglary_tot_arrests',
                 'manslaught_neg_tot_arrests', 'murder_tot_arrests', 'oth_assault_tot_arrests',
                 'oth_sex_off_tot_arrests', 'rape_tot_arrests', 'robbery_tot_arrests']
non_violent = list(set(arrest_cols).difference(set(violent_crime)))
print(len(violent_crime))
print(len(non_violent))
print(len(arrest_cols))

# EDA

merged.head()

non_violent_df = merged[['month', 'year', 'fips_state_county_code', 'population'] + non_violent].copy()
print(non_violent_df.shape)

non_violent_df['grand_total'] = np.sum(merged[non_violent], axis=1)
non_violent_df['arrest_rate_gt'] = non_violent_df['grand_total'] / non_violent_df['population'] * 100_000
non_violent_df.head()

non_violent_df['year'] = non_violent_df['year'].astype(str)
non_violent_df['day'] = "01"

non_violent_df['month'] = non_violent_df.month.str.capitalize()

non_violent_df['date'] = pd.to_datetime(non_violent_df.year +
                                        "-" + non_violent_df.month +
                                        "-" + non_violent_df.day,
                                        format="%Y-%B-%d")
non_violent_df.head()

post_treatment = "2020-June-01"
treatment_group = "08031"
non_violent_df['post_treatment'] = np.where(non_violent_df['date'] >= post_treatment, 1, 0)
non_violent_df['treatment'] = np.where(non_violent_df['fips_state_county_code'] == treatment_group, 1, 0)
non_violent_df.head()

non_violent_df = non_violent_df.query("fips_state_county_code in ['08031', '08014']")

non_violent_df

# Modeling

model = smf.ols('arrest_rate_gt ~ treatment + post_treatment + treatment * post_treatment',
                data=non_violent_df).fit()

print(model.summary())
